requirements.txt
pandas>=2.0
numpy>=1.24
openai>=1.0
anthropic>=0.30
pdfplumber>=0.10
python-docx>=1.0
openpyxl>=3.1
Pillow>=10.0
pyarrow>=14.0
pyyaml>=6.0
python-dotenv>=1.0
arize-phoenix>=8.0
openinference-instrumentation
opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk
